refactor: Rename memoswap into parity
