+++ src/components/Settings/MemoSwapParity.tsx
